feat: allow passing a custom base URL when building a project
When not hosting the project at the root of a domain or subdomain, the
paths in the `index.html` page must be adjusted to cope for the
different base.

The new argument allows to prefix all paths with a custom base URL as
well as making them relative by passing `.` as value.